Segmentation/Data processing/requirements.txt
absl-py==0.13.0
affinder==0.2.3
alphashape==1.3.1
ansicon==1.89.0
appdirs==1.4.4
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
astunparse==1.6.3
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
blessed==1.20.0
cachetools==4.2.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1707022139797/work/certifi
charset-normalizer==2.0.5
click==8.1.7
click-log==0.4.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1704278392174/work
contourpy==1.1.1
cycler==0.12.1
debugpy @ file:///C:/b/abs_c0y1fjipt2/croot/debugpy_1690906864587/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
distro==1.9.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
filelock==3.11.0
flatbuffers==1.12
flyingcircus==0.1.4.1
flyingcircus-numeric==0.1.1.4
fonttools==4.47.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.2
imageio==2.33.1
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1703269254275/work
importlib-resources==6.1.1
ipykernel @ file:///D:/bld/ipykernel_1707326433944/work
ipython @ file:///D:/bld/ipython_1680185618122/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
jinxed==1.2.1
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1699283905679/work
jupyter-core @ file:///D:/bld/jupyter_core_1669775263822/work
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy-loader==0.3
Markdown==3.3.4
matplotlib==3.7.4
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mpmath==1.3.0
napari-segment-anything==0.1.3
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.1
numpy==1.24.4
numpy-stl==3.1.1
oauthlib==3.1.1
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas==2.0.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.2.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
protobuf==3.17.3
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
pygorpho==1.0.0
pyparsing==3.1.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-utils==3.8.2
pytz==2024.1
PyWavelets==1.4.1
pywin32==227
pyzmq @ file:///C:/b/abs_89aq69t0up/croot/pyzmq_1705605705281/work
raster-geometry==0.1.4.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
Rtree==1.3.0
scikit-build==0.17.6
scikit-image==0.21.0
scipy==1.10.1
segment-anything==1.0
setuptools-scm==8.0.4
shapely==2.0.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy==1.11.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
termcolor==1.1.0
tifffile==2023.7.10
tomli==2.0.1
torch==1.9.0
torchvision==0.15.1
tornado @ file:///D:/bld/tornado_1656937958156/work
tqdm==4.62.2
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1704212992681/work
trimesh==4.4.3
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1702176139754/work
tzdata==2024.1
urllib3==1.26.6
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
